back-end/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
oauthlib==3.2.2
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
psycopg==3.1.14
psycopg-binary==3.1.14
pycparser==2.21
pypdfium2==4.24.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
